Create viewModels package && Move view model classes to it
