+++ app/src/main/java/ecb/manifest/kowalski/obd_scan/ui/presentation/bluetooth_page/BluetoothPage.kt
@@ -18,6 +18,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import ecb.manifest.kowalski.obd_scan.ui.viewModels.BluetoothPageViewModel
 
 @Composable
 fun BluetoothPage(
