+++ app/src/main/java/ecb/manifest/kowalski/obd_scan/ui/viewModels/MainActivityViewModel.kt
@@ -1,4 +1,4 @@
-package ecb.manifest.kowalski.obd_scan.ui.main
+package ecb.manifest.kowalski.obd_scan.ui.viewModels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
